requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2024.11.18
Flask==2.3.3
SQLAlchemy==2.0.25
plotly==5.18.0
pandas==2.1.4
APScheduler==3.10.4
psutil==5.9.8
